yandex_project/postgres_to_es/requirements.txt
python-dotenv==0.20.0
elastic-transport==8.1.1
elasticsearch==7.7.0
redis==4.2.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
backoff==1.11.1
psycopg2-binary==2.9.3
backoff==1.11.1